2nd review of Baking
StepDetailFragment.java

1.It is required to restore the player's play state too.

If the video is paused and the screen is rotated, the video should remain paused and vice-versa.
Currently, the paused video stops playing when the screen is rotated.
You need to save and restore mExoPlayer.getPlayWhenReady()
2.As the mExoPlayer is released here, it freezes when the app resumes from the background while the video is being played.You need to reinitialize it in onResume so that the video is played when the app resumes.You can see Demo App from the ExoPlayer team for reference.
StepDetailActivity.java
You need to put a check on savedInstanceState so that the fragments are not reinitialized twice when the screen is rotated.
This is resulting in the video being played twice when the screen is rotated while the video is being played.

You can see this for more info.

   if (savedInstanceState == null) {

            StepDetailFragment detailFragment = new StepDetailFragment();

            detailFragment.setSteps(stepData);
            detailFragment.setPosition(position);

            //Use a FragmentManager and transaction to add the fragment to the screen
            FragmentManager fragmentManager = getSupportFragmentManager();
            //Fragment transaction
            fragmentManager.beginTransaction()
                    .add(R.id.detail, detailFragment)
                    .commit();

        }